Added debug-script
Runs large and critical database operations that should never be neccesarry when not changing database-layout/structure.

When making a lot of breaks during development and a lot of redesigns, you happen to have invalid data in the database that needs to be fixed (or just delete everything. lol).
I'll now collect some of these I'm writing in case I need them in the future. Maybe replace them with Statistics etc. in the future - who knows