Optimize skin/cape history sql queries
Using `ACCESS EXCLUSIVE` slows down other unrelated queries **a lot** as SELECTs are not allowed on that table too! But we have to make sure that our SELECTs checking for existing capes/skins is not executed while anotherone is. So transaction execution time should be minimized!!

Fixes capes/skins being added multiples times (player x changes cape from id 1 to id 1 - technically not wrong as we could more presicely track when it was changed... But I honestly don't care enough to put that extra weight on my database server.... plus: We want to have a workin beta at least some time in the near future, right?)