feat: add Sentry SDK to track detailed performance metrics
I'm planning a recode of the project to finally make the projekt maintainable again.

This was my first Node.js and first TypeScript project ever and it shows.
I've learned so much since then and stopped trying to maintain this project
because the code is in a veery bad state.
When I switched servers hosting this app I realized that the project is currently
stuck with Node.js v12 and can't upgrade with some bigger updates...

So in preperation for SkinDB, this API is going to get good again!