chore: initial project setup for the re-code
I've decided to use `node --watch` instead of nodemon.

Sadly the flag is still experimental in Node.js v20 and v22 won't be LTS until the end of the year.
For the time being, we'll have to live with the warning printet :/

I've decided on essentially skipping v1 for this project, as the project is at
v0.8.0 right now and it feels wrong to just say that this recode would be v1 and not v2.