build: configure the project to output ESM instead of CommonJS
We'll try and see how well it works with this. I've read Jest, our test runner,
might have problem with ESM.
I could also imagine third-party libraries having problems.

I think having ESM output feels more modern (and less bloated).
And all the TypeScript code will be ESM, so maybe we don't have to turn it into CommonJS.

Sadly the loader flag is still experimental in Node.js v20 and v22 won't be LTS until the end of the year.
For the time being, we'll have to live with the warning printet :/