feat: implement MinecraftSkinService that fetches and normalizes a skin
This class is able to determine and return the *effective* skin for a given user.
It returns a normalized version of the skin, stores skins in the database for quicker access,
and falls back to return steve or alex steve, depending on what that user would
be presented with in the game, without having a skin.

It does not fallback to steve or alex when there is an error fetching the skin etc.