refactor: merge NotFound- and BadRequest errors into one file
Merging these two HttpErrors and upcoming ones into one
file makes the line of imports in a router smaller,
as a router might make use of multiple error classes.

Having them in one file shouldn't be an issue as the error
classes themselves are very short and do not contain any logic.