fix: refresh MaterializedView for blocked Minecraft server before update
The implementation relies on the view to contain up-to-date data.
Relying on the view, allows us to not replicate the query
used by the view in Prisma (would essentially be duplicate code).

So to make sure any potential external changes are reflected
in the materialized view, we refresh it beforehand.