feat: implement background tasks scheduler to update MC server blocklist
Mojang's API endpoint has a caching header set to
cache for a maximum of 300 seconds (5 minutes).
That's why we are running the background task every 5 minutes.

To prevent unnecessary load on the database server, I've
added a check that checks if the content has changed since the last time.
Sadly, the web server is not sending any helpful headers here :/