feat: implement Minecraft server status ping classes
The implementation is based on what the vanilla Minecraft client (v1.21) does.
It tries a "normal" ping first with the new protocol. If that fails,
it retries with a legacy ping.

I've added a 4 second delay to retry the legacy ping, as some
of the public legacy servers I tested with seem to have some kind
of protection in-place sometimes, that block the legacy ping retry,
if it follows too quickly.
That protection might not target new Minecraft clients but just
a normal "hey don't try to connect so often so fast".

In the future, these ping requests need to be proxied to
not expose the host's IP running the this app.
Maybe we can remove that delay when we rotate proxy IPs.