feat: implement legacy endpoint that renders block textures in 3D
The `modelRender.ts` file was originally created by @NudelErde.
Still many thanks!!

It has been around before the recode that I am working on
right now and looks messy etc.
But in my experimenting with WebGL and other stuff,
I failed to re-code a renderer that produces results similar to this one.

But it still contains a bug related to the front-body of a player's skin,
as it is rendered mirrored.

But for the v1 endpoints, this is good enough.