fix: always return FastifyReply from request handler
The Fastify API docs are pretty clear on how responses should be sent.
I decided on the explicit approach that calls #send on the reply etc.

To make sure faulty handler implementations to not hang
a request indefinitely or something, we always return the reply
and have the wrapping `#handleRestfully` make sure the
documented behavior is achieved.

This also fixes FastifyInstace#inject calls not working properly,
as they expect the request handler to behave in a specific way.