feat: clear expired values in SetWithTtl using the TaskScheduler
This way, we hit expired keys more often in the Sets but that's expected.
The `setTimeout` call functioned more as a sanity check
to not just have it endlessly growing memory consumption.

This also helps with our jest tests that might create such
an instance and just leaves a dangling SetWithTtl reference
with a timer running (jest warns us about un-closed references).