test: add unit tests for the ImageManipulator class
With this, the file has a statement coverage of ~93 %.
The only method that is not fully covered is `mergeColors`
which is used when rendering the (optionally) transparent Minecraft skin overlay (2nd skin layer).

I am not feeling too confident right now, to write test data
that actually makes sure, the colors are merged correctly.