ci: install needed system packages for using the `gl` npm dependency
For now, `gl` is a needed dependency which causes
`npm ci` to fail, because it can't build/install the package.

I took the list of packages to install from
https://github.com/stackgl/headless-gl/blob/254da5180eb123dd410313c66d581a98bb4266a0/.github/workflows/ci.yml

Maybe when #136 is closed or @NudelErde can help,
we can get rid of the dependency.
(It also slows down the install process by **a lot**, which is annoying).