fix: unencoded/wrongly decoded skin url fetches a somewhat broken URL
In Firefox just setting `url=https://example.com` ends up
as `https:/example` in the web server.
The URL constructor is able to still parse this and the
HttpClient can handle that too.

But for better debug logging and because we are parsing
the URL anyway for validation, we just use the 'validated'
form of the URL and everything is proper again.