fix: disable traces in Sentry SDK for now because of performance issues
I have some concerns about application performance.
It might be especially bad at app start-up but I've noticed
that the very first `UpdateMinecraftServerBlocklistTask` execution
takes over 5 seconds, which is the default transaction timeout for Prisma ORM.
But following runs of the task are fine and succeed.

Because I do not have proper performance monitoring 'from the outside' yet,
I've decided to disable it completely for now and consider
re-enabling it when it's in production when I have some kind of metric.