feat!: changed the database schema to store less duplicate skin images
I think I decided on it being an okay trade-off, I'm not sure.
But whatever my reasons were, I've decided to change the schema.
It is very possible and expected for the same "normalized" skin
to have a couple of variations that result in the same normalized skin.

This mostly saves on storage space and *might* make
direct skin lookups slightly faster while looking up the normalized skin is probably slightly slower (an additional lookup).