feat: make HttpClient abstract and introduce new clients with proxy pool
We want to use proxies to hide our server's IP in production,
as the API will be running behind Cloudflare but we accept
user-input for skin URLs and hosts to ping as Minecraft servers.

No sensitive data is transmitted, so we can use third-party proxy services.
I do intend to only use encrypted proxy communication,
but for technical reasons the proxy would be able to read the data.

The implementation etc. still needs some work and testing,
but this is the first "draft".