feat: re-implement how the HttpClient works and add proxy server support
The ProxyPoolHttpClient is supposed to help us hide
the main machines IP address, as the API is being hosted behind Cloudflare.
Cloudflare hides the machines IP and I do not want to just
expose that IP that easily via user-supplied Skin URLs or
Minecraft server IPs/Hostnames.

This also has the nice side-effect of allowing us to load-balance
the requests and use different IPs which might reduce Rate-Limit related troubles.

SOCKS(5) proxied should probably be prefered for their
simplicity and because, compared to http proxies,
https traffic is actually encrypted between the app and proxy
and not just between the proxy and the remote server.

SOCKS5 proxies also allow us to ping Minecraft servers,
which do not use HTTP.