test: skip LabyMod test for user without cape because of LabyMod changes
Right now, LabyMod returns a generic cape image for
players that do not have a cape.
I'm not sure if this is temporary or not, so I'm skipping the test case for now.

Right now, you can identify the generic skin based on the
response headers. They are different to a genuine cape response.