feat: implement writing proxy health metrics into QuestDB (time series)
This will allow me to visually see the *reference* RTT for
all the configured proxies and more easily identify troublesome ones.
It's complicated to come up with a good system that automatically
identifies and apropriatly handles proxies that are functional
but in some way not optimal while still maintaining
API operations (e.g. all proxies suddenly have 2 second referenceRTT
I don't want all of them to not be used and my API to respond with errors instead of just being slow for some time).