feat: re-implement MinecraftProfileService class and fix race condition
The main reason for recording this component was solving
the race condition of two requests for the same profile
hitting the application at (roughly) the same time and causing
two outgoing API requests to Mojang.
We do now instead wait for the already running request and
use its result, if one is already running.

This change, in theory, should only have advantages, as the
exact same code is not run twice but only run once and its result used twice.


Big thanks to @NudelErde for helping be conceptualization/design (consulting or something).