feat: track used skins for a given profile in the database
The previous version of SpraxAPI did have a history feature
that tracked the actual history of skin changes (it was able to observe).
It was able to track skin a, then b, then a again being changed.

This new system just tracks what skins have been used by a
given profile/player.

The history in the previous version only got added for SkinDB,
because other Minecraft user sites tracked skins.
But it is overly complicated and I've decided on not tracking
a full comprehensive history for every player and display it.
There's no real reason in that – I actually saw a lot of players
trying to get their "public" Minecraft data (including skin/cape history) deleted.

SkinDB will probably just display used skins in the last X days per profile.
SpraxAPI still tracks *all* used skins, it knows of, as this
information can usually be deferred from the data
that exists in other places of the database (notably texture values).