feat: store capes in the database for caching and as player cape history
The cape history is fundamentally very similar to the skin history.
The biggest difference is, that it also tracks the history of
some third-party capes, which we cannot import, unless we
freshly fetch the cape ourselves.

Mojang capes are part of a profile's texture that we can verify thanks to its signature.
LabyMod, OptiFine etc. work different in that regard, as
profiles do not need to be verified client-side (as the client is the one fetching them).