refactor: Split Prisma ORM schema files into multiple schema files
Splitting the schema definition into multiple files/groups makes
it easier to read.
As the recode got more and more complex, the schema file
grew and it becomes more and more annoying to navigate etc.
Maybe splitting into multiple files makes it easier - Let's try it ^^

This uses a preview feature of Prisma ORM:
https://www.prisma.io/docs/orm/prisma-schema/overview/location#multi-file-prisma-schema

It has been a couple of minor level releases and I think it
is worth trying out. Not sure when it will leave the preview phase.