refactor: rework how the app bootstraps the WebServer vs. Queue-Worker
These changes make the main.ts easier to read and
clearly split the individual entry point codes into their own files.

This is also in preparations to a potential CLI mode of some sorts.

The shutdown also got a bit easier and makes use of
tsyringe's Disposable interface/behavior.
Stuff we don't really need to explicitly shutdown, is handled
by disposing the container.
This has the advantage, of not using `container#resolve`,
which might create a new instance just so we can shut it down.