feat: completely rewrite how Minecraft data is written into the database
The rewrite should greatly reduce the problems I encountered
with circular dependencies (in the constructors).

It should also reduce the duplicate code a bit, as there is
one persister that persists that one relation
(two at most in the example of the SkinPersister, that also stores the URL).

There's still some refactoring potential and
missing test coverage on the new files.