test: enable TypeScript compiler option `isolatedModules`
ts-jest logged a lot of warnings when running tests about the
option needing to be enabled.

It seemed like an option that would potentially make sense
for the whole project and not just in the test tsconfig.json.

Sounded like an option that only produces warnings but
no behavior changes.
And the warnings probably warn you if you do something that,
might be weird anyways.