build: symlink python binary in the Container builder stage
node-gyp seems to detect python3 as the binary
but the gl dependency fails to call the python binary.
Not sure where the actual issue lies and it seems to be an
edge-case for target platforms etc. that have no pre-built binaries...
But we'll install the package anyways to maybe not have similar issues in the future.