fix: only run ProxyPoolHttpClientHealthcheckTask when client was used
This ensures we are not running healthchecks,
when we've never done any HTTP requests using the
proxy pool.
This might especially happen when using the CLI,
where no requests may be done.
But the CLI most definitely wants the health check *if* it
uses the proxies/HttpClient.
But also in the future, there may be different scenarios.