test: migrate the project from jest to vitest for speed and compatibility
vitest feels pretty nice. Is mostly compatible with jest
which made the migration of the unit tests pretty easy.
It is fast and has a nice watcher.
It supports my ESM and TypeScript code out-of-the-box,
without any issues.
(And I started having more issued with jest, some of which
I already worked around in the past).

**But** because it is using esbuild under-the-hood
without any way to configure this (and I tried a couple of workarounds),
tsyringe does not work, as the typescript option
`emitDecoratorMetadata` is not supported.
This is why the `vitest.setup.ts` is soooo big.
The dependency container doesn't work anymore, that
the acceptance tests rely on.
Manually (and only partially) populating it is not the
idea of the acceptance tests, but I couldn't find any better
way to do this.
My biggest hope is for decorators to become JS native in
the future or esbuild to support them (maybe with the
tsc rewrite to Golang?) or vitest allowing us to use swc or tsc.
My next possibility is rewriting the tests to start the actual
app and have the tests use that instead. Which would more
correctly reflect the actual app as it is normally run, instead
manually and pragmatically injecting a request.