fix: allow multiple instances of the container image on the same network
When a screen/x-server/dunno already exists on the
network the container is using (e.g. two instances of SpraxAPI),
then the second container is unable to start, because the
server number/ID is already taken.
We can instruct xvfb-run to automatically choose one for us,
instead of using the default one, which might already be taken