feat: fall back to a third-party API for username to UUID resolution
The Mojang API can sometimes be unreliable, and I guess
there is no official documentation (and it is not a public API).
The username to UUID endpoint is especially problematic,
often returning a status 403 from Azure.

For some improved stability, we now try to resolve the UUID
using mcproxy.dev (by @InventivetalentDev)
and fetch the profile from Mojang afterwards, just to
make sure the resolved username is actually valid.
Mojang's profile API is more reliable, so hopefully this
solves the issue for us (at least most of the time).