fix: have SkinPersister acquire a lock on the skin URL when needed
Because the queue is worked on concurrently,
race conditions may occur, depending on the queued items
and the order they are worked on.

Locking on a skinURL that might not exist yet helps here,
because the 'existing' check might return 'does not exist yet'
while another queue item creates it and when we
finally want to persist the SkinURL we suddenly conflict
with the unique constraint, because by now the URL does exist.